pyproject.toml
[tool.poetry]
name = "volue.mesh"
version = "0.1.0"
description = ""
license = "Proprietary"
authors = ["Volue AS <erik.nyhus@volue.com>"]
readme = "README.md"
homepage = "https://github.com/PowelAS/sme-mesh-python"
repository = "https://github.com/PowelAS/sme-mesh-python"
packages = [
    { include = "volue/mesh", from = "src" }
]
# By default Poetry doesn't package files in .gitignore.
include = [
    "src/volue/mesh/proto/mesh_pb2.py",
    "src/volue/mesh/proto/mesh_pb2_grpc.py"
]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "1.37.0"
pyarrow = "^4.0.1"
protobuf = "^3.17.3"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.37.0"
Sphinx = "^4.0.2"
m2r2 = "^0.2.1"

[build-system]
requires = [
    "poetry-core>=1.0.3",
    "grpcio-tools>=1.37.0"
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
script = "build.py"
generate-setup-file = false
